added container components for all metrics, separating concerns and cleaning up dashboard
