+++ .eslintignore
@@ -3,3 +3,4 @@
 **/*.min.js
 **/coverage/*
 **/build/*
+**/docs/*
