+++ .eslintrc.json
@@ -13,14 +13,17 @@
       "next": true
     },
     "extends": "eslint:recommended",
+    "parser": "babel-eslint",
     "parserOptions": {
-      "sourceType": "module"
+      "sourceType": "module",
+      "ecmaVersion": 2018
     },
     "rules": {
       "no-console": "off",
       "indent": [ "error", 2 ],
       "quotes": ["error", "single", { "allowTemplateLiterals": true }],
       "comma-dangle": ["error", "always-multiline"],
-      "semi": [ "error", "always" ]
+      "semi": [ "error", "always" ],
+      "no-unused-vars": [1, { "vars": "local", "args": "none" }]
     }
   }
