+++ .travis.yml
@@ -1,0 +1,10 @@
+language: node_js
+node_js:
+  - "stable"
+branches:
+  only:
+    - master
+script:
+  - npm run lint
+  - npm test
+
