+++ src/components/cost/cost-item.js
@@ -1,11 +1,11 @@
-import React, { Component, Fragment } from 'react';
+import React from 'react';
 import CostForm from './cost-form';
-export default class CostItem extends Component {
+export default class CostItem extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
       view: false,
-    }
+    };
   }
 
   updateView = () => {
@@ -15,40 +15,36 @@ export default class CostItem extends Component {
     this.setState({view: false});
   }
   onRemove = () => {
-      this.props.onRemove(this.props.cost);
-
-    }
-  
- costUpdate = (cost) => {
-  this.props.onComplete(cost);
-
-
+    this.props.onRemove(this.props.cost);
+  }
+  costUpdate = (cost) => {
+    this.props.onComplete(cost);
   }
 
   render() {
     let {cost: {station, task}} = this.props;
-      return(
-        <Fragment>
-         <div onDoubleClick={this.updateView}>
-            <div className="list-wrapper">
-              <div className="row">
-                <div className="column">
-                  <div className="column1">
-                    <h3>STATION</h3>
-                    <p>{station}</p>
-                  </div>
+    return(
+      <div>
+        <div onDoubleClick={this.updateView}>
+          <div className="list-wrapper">
+            <div className="row">
+              <div className="column">
+                <div className="column1">
+                  <h3>STATION</h3>
+                  <p>{station}</p>
                 </div>
-                <div className="column">
-                  <div className="column2">
-                    <h3>TASK</h3>
-                    <p>{task}</p>
-                  </div>
+              </div>
+              <div className="column">
+                <div className="column2">
+                  <h3>TASK</h3>
+                  <p>{task}</p>
                 </div>
               </div>
             </div>
-            {this.state.view && <button onClick={this.onRemove}>Delete</button>}
-         </div> 
-         {this.state.view && 
+          </div>
+          {this.state.view && <button onClick={this.onRemove}>Delete</button>}
+        </div> 
+        {this.state.view && 
            <div>
              <CostForm 
                onComplete={this.costUpdate} 
@@ -58,8 +54,8 @@ export default class CostItem extends Component {
                buttonText = 'save update'
              />
              <button onClick={this.returnView}>cancel update</button>
-          </div>}
-        </Fragment>
-      );
+           </div>}
+      </div>
+    );
   }
 }
