+++ src/components/quality/quality-item.js
