+++ src/components/rollup/rollup-form.js
@@ -1,0 +1,61 @@
+import React from 'react';
+import dateFormat from 'dateformat';
+const now = new Date();
+export default class RollupForm extends React.Component {
+  constructor(props) {
+    super(props);
+    this.defaultState = {
+      tVarience: '',
+      tCompleted: '',
+      endOfDay: '',
+      areaVarience: '',
+      date: dateFormat(now,'mmm dd, yyyy'),
+    };
+    const initialState = this.props.rollup || this.defaultState;
+
+    this.state = {...initialState};
+  }
+  // setDate = (e) => {
+  //     e.preventDefault();
+  //     let day = new Date(e);
+  //     this.setState({date: day.toLocaleDateString("en-US")});
+  // }
+
+  onSubmit = (e) => {
+    e.preventDefault();
+    this.props.onComplete(this.state);
+    this.props.viewChange();
+    this.setState({...this.defaultState});
+  };
+  handleChange = (e) => {
+    const { name, value } = e.target;
+    this.setState({ [name]: value });
+  };
+
+  render() {
+    return(
+      <div>
+        <div class="form-wrapper">
+          <fieldset>
+            <form onSubmit={this.onSubmit} onChange={this.handleChange}>
+              <label>Total Varience
+                <input name="tVarience" type="text" value={this.state.tVarience}/>
+              </label>
+              <label>Total Completed
+                <input name="tCompleted" type="text" value={this.state.tCompleted}/>
+              </label>
+              <label>End Of Day
+                <input name="endOfDay" type="text" value={this.state.endOfDay}/>
+              </label>
+              <label>Area Varience
+                <input name="areaVarience" type="text" value = {this.state.areaVarience}/>
+              </label>
+              <button>{this.props.buttonText}</button>
+            </form>
+          </fieldset>
+        </div>
+      </div>
+    );
+  
+  }
+}
