+++ src/components/safety/safety-item.js
