+++ src/redux/store.js
@@ -5,7 +5,7 @@ import safetyReducer from './safety/safety.reducer';
 import qualityReducer from './quality/quality.reducer';
 import deliveryReducer from './delivery/delivery.reducer';
 import costReducer from './cost/cost.reducer';
-import moraleReducer from './morale/morale.educer';
+import moraleReducer from './morale/morale.reducer';
 import rollupReducer from './rollup/rollup.reducer';
 
 const appReducer = combineReducers({safety: safetyReducer, quality: qualityReducer, delivery: deliveryReducer, cost: costReducer, morale: moraleReducer, rollup: rollupReducer});
